ADD: Auth route and middleware

+++ NLW12-spacetime/server/src/routes/memories.ts
@@ -3,8 +3,17 @@ import { prisma } from '../lib/prisma'
 import z from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        user: {
+          id: req.user.sub,
+        },
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -32,6 +41,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    if (!memory.isPublic && memory.userId !== req.user.sub)
+      return res.status(401).send()
+
     res.send(memory)
   })
 
@@ -51,7 +63,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'b71dcb1f-f646-4a19-9369-1f367c8247ae',
+        userId: req.user.sub,
       },
     })
 
@@ -72,7 +84,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(req.params)
     const { content, coverUrl, isPublic } = memorySchema.parse(req.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) return res.status(401).send()
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -92,6 +112,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) return res.status(401).send()
 
     await prisma.memory.delete({
       where: {
